Add files via upload
In order to demonstrate fundamental programming ideas like inheritance, polymorphism, and event-driven design, this project focusses on creating a contemporary version of the traditional Snake game in C++ using the SDL2 library. The user controls a snake in the game that travels across the screen and eats food, getting longer with each bite. Avoiding collisions with the snake's body or the screen's edges is the difficult part because doing so will end the game. The game manages user inputs, graphical rendering, and real-time updates using SDL2, offering a responsive and aesthetically pleasing gaming experience. The game becomes modular, reusable, and expandable by organising the project according to object-oriented programming concepts, which enables additional customisation or the addition of new features.